requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.10
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-filter==2.3.0
django-js-asset==1.2.2
filelock==3.0.12
httplib2==0.9.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
nemo-emblems==4.4.0
netifaces==0.10.4
onboard==1.4.1
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparted==3.11.1
pyRFC3339==1.0
python-apt==1.6.4
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
tinycss==0.4
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.10
virtualenv==20.0.31
waitress==1.4.4
whitenoise==5.2.0
xkit==0.0.0
zipp==3.2.0
